organize module directory
