+++ src-deprecated/AppMetaModule.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BEAR\Package;
+
+use BEAR\Package\Module\AppMetaModule as AppMetaModuleInModule;
+
+/**
+ * @deprecated use BEAR\Package\Module\AppMetaModule instead
+ */
+class AppMetaModule extends AppMetaModuleInModule
+{
+}
