+++ src/Module.php
@@ -7,7 +7,7 @@ namespace BEAR\Package;
 use BEAR\AppMeta\AbstractAppMeta;
 use BEAR\Package\Exception\InvalidContextException;
 use BEAR\Package\Exception\InvalidModuleException;
-use BEAR\Package\Provide\Boot\CacheNamespaceModule;
+use BEAR\Package\Module\CacheNamespaceModule;
 use Ray\Di\AbstractModule;
 use Ray\Di\AssistedModule;
