+++ src/Module/ScriptinjectorModule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace BEAR\Package\Provide\Boot;
+namespace BEAR\Package\Module;
 
 use Ray\Compiler\ScriptInjector;
 use Ray\Di\AbstractModule;
