+++ src/PackageModule.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace BEAR\Package;
 
-use BEAR\Package\Provide\Boot\DiCompileModule;
+use BEAR\Package\Module\DiCompileModule;
 use BEAR\Package\Provide\Error\VndErrorModule;
 use BEAR\Package\Provide\Logger\PsrLoggerModule;
 use BEAR\Package\Provide\Representation\CreatedResourceModule;
